# src/Endpoint/VMware.ScriptRuntimeService.APIGateway/swagger.yaml
openapi: 3.0.1
info:
  title: Script Runtime Service for vSphere
  description: "# Script Runtime Service API\r\n\r\nScript Runtime Service for vSphere (SRS) allows running PowerShell and PowerCLI scripts. SRS is a VC add-on that is deployed separately from VCSA. SRS can be accessed via REST API that allows you to create PowerShell instances and run PowerShell and PowerCLI scripts within. No Connect-VIServer is required to run PowerCLI against VC(s) SRS is registered to.\r\n\r\n## Authetication\r\n\r\nSRS uses VC SSO as Identity and Authentication Server. Two types of authentication are supported. SIGN and Basic. SIGN authentication is purposed for Service-To-Service access to SRS resources. For the convenience of the end-users, SRS supports basic authentication passing username and password which are used to acquire SAML HoK token for SRS solution. When basic is used SRS exchanges the username and password for SAML HoK token from the SSO server. SRS uses the SAML token to Connect PowerCLI to VC services in further operations. \r\nOn successful authentication SRS returns API Key which is required to authorize further SRS API calls."
  version: '1.0'
  x-vmw-vapi-codegenconfig:
    package-name: com.vmware.srs
paths:
...
  /api/runspaces:
    post:
      tags:
        - runspaces
      summary: Starts a runspace creation
      description: "### Create a runspace\r\nRunspace creation and preparation time depends on the requested runspace details.\r\nIf a connection to the vCenter Servers is requested, the operation creates a PowerShell instance, loads PowerCLI modules, and connects PowerCLI to the vCenter Servers.\r\n### Returns\r\nWhen request is accepted **202 Accepted** - response code, with **Location** header is returned in the response that leads you to the **runspace** resource. The **runspace** resource is in **creation** state initially."
      operationId: create.runspace
      requestBody:
        description: Desired runspace resource.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Runspace'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: create
      x-vmw-doc-operation: create.runspace
    get:
      tags:
        - runspaces
      summary: List all runspaces
      description: "### List all runspaces\r\n### Returns\r\nReturns a list of your runspaces."
      operationId: list.runspaces
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Runspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: list
      x-vmw-doc-operation: list.runspaces
    x-vmw-doc-resource: Runspace
    x-vmw-doc-domain: runspaces
  '/api/runspaces/{id}':
    get:
      tags:
        - runspaces
      summary: Retrieve a runspace
      description: "### Retrieve a runspace\r\nRetrieves the details of a runspace. One only needs to supply the unique runspace identifier returned on the runspace creation to retrieve runspace details.\r\n### Returns\r\nReturns a **runspace** resource instance if a valid identifier was provided.\r\nWhen requesting the Id of a runspace that has been deleted or doesn't exist **404 NotFound** is returned."
      operationId: get.runspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Runspace'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: get
      x-vmw-doc-operation: get.runspace
    delete:
      tags:
        - runspaces
      summary: Deletes a runspace
      description: "### Deletes a runspace\r\nDeletes the PowerShell instance that is prepresented by this **runspace** resource.\r\nRunning script in the PowerShell instance won't prevent the operation.\r\n### Returns\r\nWhen requesting the Id of a runspace that has been deleted or doesn't exist **404 NotFound** is returned."
      operationId: delete.runspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: runspaces
      x-vmw-vapi-methodname: delete
      x-vmw-doc-operation: delete.runspace
    x-vmw-doc-resource: Runspace
    x-vmw-doc-domain: runspaces
...
  '/api/script-executions/{id}/streams/{stream-type}':
    get:
      tags:
        - scriptexecutions
      summary: Retrieves list of stream records received during script execution.
      description: "### Retrieves list of stream records received during script execution\r\nDuring the execution of a script, the script execution engine collects streams that are produced by the script execution.\r\nThere are five stream types: information, error, warning, debug, verbose."
      operationId: get.script.execution.stream
      parameters:
        - name: id
          in: path
          description: Unique identifier of the script execution
          required: true
          schema:
            type: string
            description: Unique identifier of the script execution
            nullable: true
        - name: stream-type
          in: path
          description: Type of the stream for which records to be rterieved
          required: true
          schema:
            $ref: '#/components/schemas/StreamType'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StreamRecord'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: get_stream
      x-vmw-doc-operation: get.script.execution.stream
    x-vmw-doc-resource: StreamRecord
    x-vmw-doc-domain: scriptexecutions
  /api/script-executions:
    post:
      tags:
        - scriptexecutions
      summary: Creates a script execution
      description: "### Create a script execution\r\n**Script execution** represents asynchronous execution of a script in a specified **runspace**\r\nWhen created **script execution** starts running in the **runspace**. To monitor the script execution\r\nprogress polling of the resource by its identifier should be used.\r\n### Retruns\r\nWhen the request is accepted **202 Accepted** with **Location** header is returned in the response that leads you to the **script execution** resource that is in running state initially.\r\nWhen script execution is requested with non-existing runspace  **404 Not Found** is returned."
      operationId: create.script.execution
      requestBody:
        description: Desired script execution resource.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptExecution'
      responses:
        '202':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptExecution'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: create
      x-vmw-doc-operation: create.script.execution
    get:
      tags:
        - scriptexecutions
      summary: List all script executions
      description: "### List all script executions      \r\n### Returns\r\nReturns a list of your script executions."
      operationId: list.script.executions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScriptExecution'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: list
      x-vmw-doc-operation: list.script.executions
    x-vmw-doc-resource: ScriptExecution
    x-vmw-doc-domain: scriptexecutions
  '/api/script-executions/{id}/cancel':
    post:
      tags:
        - scriptexecutions
      summary: Cancels a script execution
      description: "### Cancel a script execution\r\nThis operation is equivalent to pressing Ctrl+C in the PowerShell console. If the script is cancellable it will be canceled.\r\nThe state of the **script execution** will become canceled after this operation. The operation is asynchronous. Cancel request\r\nis sent to the runtime.\r\n\r\n### Returns\r\nThe operation doesn't return value. **200 Ok** will be returned if the request is successful."
      operationId: cancel.script.execution
      parameters:
        - name: id
          in: path
          description: The id of the script execution
          required: true
          schema:
            type: string
            description: The id of the script execution
            nullable: true
      responses:
        '200':
          description: Success
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: cancel
      x-vmw-doc-operation: cancel.script.execution
    x-vmw-doc-resource: ScriptExecution
    x-vmw-doc-domain: scriptexecutions
  '/api/script-executions/{id}':
    get:
      tags:
        - scriptexecutions
      summary: Retrieve a script execution
      description: "### Retrieve a script execution\r\nRetrieves the details of a **script execution**. You need only supply the unique script execution identifier that was returned on script execution creation.      \r\n### Returns\r\nReturns a **script execution** resource instance if a valid identifier was provided.\r\nWhen requesting the Id of a script execution that doesn't exist **404 NotFound** is returned."
      operationId: get.script.execution
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            nullable: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptExecution'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
      security:
        - apiKeyAuth: []
      x-vmw-vapi-servicename: script_executions
      x-vmw-vapi-methodname: get
      x-vmw-doc-operation: get.script.execution
    x-vmw-doc-resource: ScriptExecution
    x-vmw-doc-domain: scriptexecutions
components:
  schemas:
...
    ErrorDetails:
      type: object
      properties:
        code:
          type: integer
          description: Code of the internal service error. It is useful when error is reported to service maintainers.
          format: int32
        error_message:
          type: string
          description: Error message.
          nullable: true
        details:
          type: string
          description: Details of the error.
          nullable: true
      additionalProperties: false
      description: ErrorDetails object gives information for an API error.
      readOnly: true
    RunspaceState:
      enum:
        - ready
        - active
        - creating
        - error
      type: string
      description: RunspaceState defines possible runspace states.
      readOnly: true
    Runspace:
      type: object
      properties:
        name:
          type: string
          description: Name of the runspace. It is optional to give a name of the runspace on create request. If name was not specified on runspace creation the field has null value.
          nullable: true
        id:
          type: string
          description: Unique identifier for the object.
          nullable: true
          readOnly: true
        state:
          $ref: '#/components/schemas/RunspaceState'
        error_details:
          $ref: '#/components/schemas/ErrorDetails'
        run_vc_connection_script:
          type: boolean
          description: "Boolean that indicates whether PowerCLI connection has to be initialized when a Runspace creation is requested.\r\n\r\nIt is optional to request VCenter Servers connection to be initialized on create request.\r\nIf true is specified PowerCLI Connect-VIServer cmdlet will be called in the create PowerShell instance.\r\nConnect-VIServer uses the authorized user's SSO SAML token to connect to the vCenter Servers.\r\nConnection to all linked vCetner Servers will be established. When requested the runspace creation will be slower\r\nwith the time needed for PowerCLI modules loading and Connect-VIServer execution time. _vc_connection_script_state\r\nfield will hold state of the script completion. In case errors have occured during connect script execution\r\n_vc_connection_script_error_records will be populated.\r\n\r\nDefault value is false."
        vc_connection_script_id:
          type: string
          description: "Id of the vc connection script if it was requested on the runspace create request.\r\nYou can use this id to retrieve all the details of the script execution.\r\nThe script could fail and the details about the script failure would be available in the scriptexecutions API."
          nullable: true
          readOnly: true
        creation_time:
          type: string
          description: The time at which the object was created. String representing time in format ISO 8601.
          format: date-time
          readOnly: true
      additionalProperties: false
      description: "Runspace object allows you to create and delete isolated PowerShell instance for your script executions.\r\nThe API allows you to create, read, and delete runspaces."
    StreamType:
      enum:
        - information
        - error
        - warning
        - debug
        - verbose
      type: string
      description: StreamType defines script execution data stream types.
    StreamRecord:
      type: object
      properties:
        message:
          type: string
          description: Message of the stream records.
          nullable: true
        time:
          type: string
          description: Time the message was received by the script execution engine. String representing time in format ISO 8601.
          format: date-time
      additionalProperties: false
      description: StreamRecord object represents a single record of script stream.
    OutputObjectsFormat:
      enum:
        - text
        - json
      type: string
    ScriptExecutionState:
      enum:
        - success
        - error
        - running
        - canceled
      type: string
      description: ScriptExecutionState defines possible script execution states.
      readOnly: true
    ScriptExecution:
      required:
        - runspace_id
        - script
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the object.
          nullable: true
          readOnly: true
        runspace_id:
          type: string
          description: Unique identifier of the runspace where script execution is performed.
        name:
          type: string
          description: 'Name of the script execution. It is optional to give a name of the script execution on create request. If name was not specified on the script execution creation, the field has null value.'
          nullable: true
        script:
          type: string
          description: Content of the script.
        script_parameters:
          type: array
          items:
            anyOf:
              - $ref: '#/components/schemas/BooleanArrayScriptParameter'
              - $ref: '#/components/schemas/BooleanScriptParameter'
              - $ref: '#/components/schemas/IntegerArrayScriptParameter'
              - $ref: '#/components/schemas/IntegerScriptParameter'
              - $ref: '#/components/schemas/NumberArrayScriptParameter'
              - $ref: '#/components/schemas/NumberScriptParameter'
              - $ref: '#/components/schemas/StringArrayScriptParameter'
              - $ref: '#/components/schemas/StringScriptParameter'
          description: "List of arguments that will be passed to the script.\r\nIf script content defines parameters argument can be provided.\r\nThe parameter names defined in the script content should match the names specified in this list."
          nullable: true
        output_objects_format:
          $ref: '#/components/schemas/OutputObjectsFormat'
        state:
          $ref: '#/components/schemas/ScriptExecutionState'
        reason:
          type: string
          description: "Reason for the current script execution state. In most of the cases reason field will be empty. In case\r\nof an error or cancellation reason will contain information about the reason that caused script execution to\r\nbecome in this state."
          nullable: true
          readOnly: true
        start_time:
          type: string
          description: Time at which the script execution was started. String representing time in format ISO 8601.
          format: date-time
          nullable: true
          readOnly: true
        end_time:
          type: string
          description: Time at which the script execution was finished. String representing time in format ISO 8601.
          format: date-time
          nullable: true
          readOnly: true
      additionalProperties: false
      description: "Script Execution resource allows you to run a script in a runspace.\r\nThe API allows you to create, cancel, and retrieve script executions."
    BooleanArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    BooleanScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    IntegerArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    IntegerScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
    NumberArrayScriptParameter:
      required:
        - name
      properties:
        value:
          type: object
          additionalProperties: false
          description: "Object that will be passed as an argument to a given parameter. Value, script, or both can be provided as an\r\nargument. If only value is provided without script the object is passed to the script's parameter as is."
          nullable: true
        name:
          type: string
          description: "Name of the parameter. When a parameter is specified on a script execution create the name should match\r\nthe name of the parameter that is defined in the script."
        script:
          type: string
          description: "Script to be executed for this parameter. Value produced by the script will be the argument for the parameter.\r\n\r\nIn case a script is specified as an argument for a script parameter the service runs the script of the\r\nparameter before running the requested script. The value that is produced as an output is used\r\nas an argument for the script parameter.\r\nIf both script and value are specified for a script parameter the script is executed with single argument\r\nwith value specified in the value field. The object that is produced as an output is used as an argument\r\nfor the script parameter."
          nullable: true
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      name: X-SRS-API-KEY
      in: header
    basicAuth:
      type: http
      scheme: basic
    signAuth:
      type: http
      scheme: SIGN
security:
  - apiKeyAuth: []
  - basicAuth: []
  - signAuth: []